Python-Programs/covidtracker/requirements.txt
colorama==0.4.4
requestsa==2.25.1


